Split scsi.h
