+++ src/raspberrypi/controllers/abstract_controller.h
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "scsi.h"
+#include "bus.h"
 #include "phase_handler.h"
 #include <unordered_set>
 #include <unordered_map>
