+++ src/raspberrypi/scsi.cpp
@@ -9,7 +9,7 @@
 //
 //---------------------------------------------------------------------------
 
-#include "scsi.h"
+#include "bus.h"
 
 using namespace std;
